+++ ui/src/router/index.tsx
@@ -14,7 +14,7 @@ import PoolListPage from '../pages/PoolListPage'
 import RateLimitPage from '../pages/RateLimitPage'
 import RuntimeConfigPage from '../pages/RuntimeConfigPage'
 import SandboxesPage from '../pages/SandboxesPage'
-import SandboxTemplateConfigPage from '../pages/SandboxTemplateConfigPage'
+import SandboxBlueprintConfigPage from '../pages/SandboxBlueprintConfigPage'
 import TemplatesConfigPage from '../pages/TemplatesConfigPage'
 import TerminalPage from '../pages/TerminalPage'
 
@@ -97,8 +97,8 @@ export const appRouter = createHashRouter([
             element: <TemplatesConfigPage />,
           },
           {
-            path: 'config/sandbox-template',
-            element: <SandboxTemplateConfigPage />,
+            path: 'config/blueprint',
+            element: <SandboxBlueprintConfigPage />,
           },
           {
             path: 'config/runtime',